requirements.txt
fonttools
brotli
zopfli
flask
pywebview
pyinstaller
